requirements.txt
# FATI Backend - Requirements
# Django REST API for Fond d'Analyse Territoriale Intégrée

# Django Core
django>=4.2.0,<5.0.0
django-cors-headers>=4.3.0
django-filter>=23.5.0

# REST Framework
djangorestframework>=3.14.0
djangorestframework-gis>=1.0

# Database
psycopg2-binary>=2.9.9

# GeoDjango
GDAL>=3.0.0

# API Documentation
drf-yasg>=1.21.0

# Authentication
PyJWT>=2.8.0

# Utilities
Pillow>=10.1.0
python-dateutil>=2.8.2

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development (optional)
ipython>=8.18.0

# Environment
django-environ>=0.11.0
